pyproject.toml
[project]
name = "2D-Game-Asset-Generator"
version = "0.1.0"
description = "AI-powered 2D game asset generator for characters, backgrounds, and items"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.33.0",
    "gradio>=5.44.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
]
